feat: devex improvements to noirjs
