+++ tooling/noir_js/src/index.ts
@@ -1,9 +1,34 @@
-import * as acvm from '@noir-lang/acvm_js';
-import * as abi from '@noir-lang/noirc_abi';
+/* eslint-disable  @typescript-eslint/no-explicit-any */
+import { Backend, ProofData } from '@noir-lang/types';
+import { generateWitness } from './witness_generation.js';
+import initAbi from '@noir-lang/noirc_abi';
+import initACVM from '@noir-lang/acvm_js';
 
-export { acvm, abi };
+export class Noir {
+  constructor(private backend: Backend) {}
 
-export { generateWitness } from './witness_generation.js';
-export { acirToUint8Array, witnessMapToUint8Array } from './serialize.js';
+  async init(): Promise<void> {
+    // If these are available, then we are in the
+    // web environment. For the node environment, this
+    // is a no-op.
+    if (typeof initAbi === 'function') {
+      await Promise.all([initAbi(), initACVM()]);
+    }
+    await this.backend.instantiate();
+  }
 
-export { Noir } from './program.js';
+  async destroy(): Promise<void> {
+    await this.backend.destroy();
+  }
+
+  // Initial inputs to your program
+  async generateFinalProof(inputs: any): Promise<ProofData> {
+    await this.init();
+    const serializedWitness = await generateWitness(this.backend.circuit, inputs);
+    return this.backend.generateFinalProof(serializedWitness);
+  }
+
+  async verifyFinalProof(proofData: ProofData): Promise<boolean> {
+    return this.backend.verifyFinalProof(proofData);
+  }
+}
