+++ tooling/noir_js_types/.gitignore
